requirements.txt
# pip install -U -r requirements.txt
requests
requests_oauthlib
psycopg2-binary
peewee>=2.10.0,<3.0.0
peewee-validates
python-slugify
emails
wtf-peewee==0.2.6
pycountry
html2text
tablib
openpyxl>=2.4.9,<2.5.0
pyyaml
pykwalify
Flask
#flask-admin
# NB! prefix with a single space for pbr!
 git+https://github.com/nad2000/flask-admin/#egg=flask_admin
flask-login
Flask-WTF
flask-peewee==0.6.7
Flask-OAuthlib
raven[flask]
flask-peewee-swagger>=1.1.5
flask-limiter>=1.0.0,<2.0.0
flask-swagger
flask-restful
jsonschema>=2.6.0,<3.0.0
validators
croniter==0.3.20
rq==0.10.0
rq-scheduler
Flask-RQ2[cli]==18.0
rq-dashboard
